Refactor CinemachineCameraManagerBase to be more useful for custom managers
