+++ com.unity.cinemachine/Runtime/Core/CinemachineCameraManagerBase.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.Serialization;
 
 namespace Unity.Cinemachine
 {
@@ -34,10 +35,24 @@ namespace Unity.Cinemachine
         [FoldoutWithEnabledButton]
         public DefaultTargetSettings DefaultTarget;
 
-        /// State for CreateActiveBlend().  Used in the case of backing out of a blend in progress.
+        /// <summary>
+        /// The blend which is used if you don't explicitly define a blend between two Virtual Camera children.
+        /// </summary>
+        [Tooltip("The blend which is used if you don't explicitly define a blend between two Virtual Camera children")]
+        [FormerlySerializedAs("m_DefaultBlend")]
+        public CinemachineBlendDefinition DefaultBlend = new (CinemachineBlendDefinition.Styles.EaseInOut, 0.5f);
+
+        /// <summary>
+        /// This is the asset which contains custom settings for specific child blends.
+        /// </summary>
+        [Tooltip("This is the asset which contains custom settings for specific child blends")]
+        [FormerlySerializedAs("m_CustomBlends")]
+        public CinemachineBlenderSettings CustomBlends = null;
+
         List<CinemachineVirtualCameraBase> m_ChildCameras;
         readonly BlendManager m_BlendManager = new ();
         CameraState m_State = CameraState.Default;
+        ICinemachineCamera m_TransitioningFrom;
 
         /// <summary>Reset the component to default values.</summary>
         protected virtual void Reset()
@@ -139,6 +154,62 @@ namespace Unity.Cinemachine
             }
         }
 
+        /// <summary>Internal use only.  Do not call this method.
+        /// Called by CinemachineCore at designated update time
+        /// so the vcam can position itself and track its targets.  This implementation
+        /// updates all the children, chooses the best one, and implements any required blending.</summary>
+        /// <param name="worldUp">Default world Up, set by the CinemachineBrain</param>
+        /// <param name="deltaTime">Delta time for time-based effects (ignore if less than or equal to 0)</param>
+        public override void InternalUpdateCameraState(Vector3 worldUp, float deltaTime)
+        {
+            UpdateCameraCache();
+            if (!PreviousStateIsValid)
+                ResetLiveChild();
+
+            // Choose the best camera - auto-activate it if it's inactive
+            var best = ChooseCurrentCamera(worldUp, deltaTime);
+            if (best != null && !best.gameObject.activeInHierarchy)
+            {
+                best.gameObject.SetActive(true);
+                best.UpdateCameraState(worldUp, deltaTime);
+            }
+            SetLiveChild(best, worldUp, deltaTime, LookupBlend);
+
+            // Special case to handle being called from OnTransitionFromCamera() - GML todo: fix this
+            if (m_TransitioningFrom != null && !IsBlending && LiveChild != null)
+            {
+                LiveChild.OnCameraActivated(new ICinemachineCamera.ActivationEventParams
+                {
+                    Origin = this,
+                    OutgoingCamera = m_TransitioningFrom,
+                    IncomingCamera = LiveChild,
+                    IsCut = false,
+                    WorldUp = worldUp, 
+                    DeltaTime = deltaTime
+                });
+            }
+
+            FinalizeCameraState(deltaTime);
+            m_TransitioningFrom = null;
+            PreviousStateIsValid = true;
+        }
+        
+        /// <summary>Find a blend curve for blending from one child camera to another.</summary>
+        /// <param name="outgoing">The camera we're blending from.</param>
+        /// <param name="incoming">The camera we're blending to.</param>
+        /// <returns>The blend to use for this camera transition.</returns>
+        protected virtual CinemachineBlendDefinition LookupBlend(ICinemachineCamera outgoing, ICinemachineCamera incoming)
+            => CinemachineBlenderSettings.LookupBlend(outgoing, incoming, DefaultBlend, CustomBlends, this);
+            
+        /// <summary>
+        /// Choose the appropriate current camera from among the ChildCameras, based on current state.
+        /// If the returned camera is different from the current camera, an appropriate transition will be made.
+        /// </summary>
+        /// <param name="worldUp">Default world Up, set by the CinemachineBrain</param>
+        /// <param name="deltaTime">Delta time for time-based effects (ignore if less than or equal to 0)</param>
+        /// <returns>The current child camera that should be active. Must be present in ChildCameras.</returns>
+        protected abstract CinemachineVirtualCameraBase ChooseCurrentCamera(Vector3 worldUp, float deltaTime);
+        
         /// <summary>This is called to notify the vcam that a target got warped,
         /// so that the vcam can update its internal state to make the camera
         /// also warp seamlessly.</summary>
@@ -173,7 +244,9 @@ namespace Unity.Cinemachine
             ICinemachineCamera fromCam, Vector3 worldUp, float deltaTime)
         {
             base.OnTransitionFromCamera(fromCam, worldUp, deltaTime);
+            m_TransitioningFrom  = fromCam;
             InvokeOnTransitionInExtensions(fromCam, worldUp, deltaTime);
+            InternalUpdateCameraState(worldUp, deltaTime);
         }
 
         /// <summary>Force a rebuild of the child camera cache.
